sw/host/opentitanlib: Refactor monolith into sub-crates and split unit tests
Refactors the large opentitanlib crate into smaller, logical sub-crates: core, chip, debug, app, image, protocols, backend, test_utils.

Resolves circular dependencies by moving core interfaces/errors to io, and moving Params instantiation to TransportWrapper.

Splits the monolithic unit tests into corresponding sub-crate tests.

TAG=agy

CONV=54fea5d4-a787-4b68-b90e-2ea68dc2baa6
