+++ sw/host/opentitanlib/src/backend_lib.rs
@@ -1,0 +1,12 @@
+// Copyright lowRISC contributors (OpenTitan project).
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod backend;
+
+// Re-export core modules for internal use
+pub(crate) use opentitanlib_core::transport;
+pub(crate) use opentitanlib_core::util;
+
+// Re-export app modules for internal use
+pub(crate) use opentitanlib_app::app;
