+++ sw/host/opentitanlib/src/core_lib.rs
@@ -2,5 +2,10 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-pub use opentitanlib_app::transport::common::fpga;
-pub use opentitanlib_core::transport::common::usb;
+pub mod crypto;
+pub mod io;
+#[path = "transport_core.rs"]
+pub mod transport;
+pub mod util;
+
+pub use ot_hal::with_unknown;
