+++ sw/host/opentitanlib/src/debug_lib.rs
@@ -1,0 +1,11 @@
+// Copyright lowRISC contributors (OpenTitan project).
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod debug;
+
+// Re-export core modules for internal use to avoid changing crate:: imports
+pub(crate) use opentitanlib_core::impl_serializable_error;
+pub(crate) use opentitanlib_core::io;
+pub(crate) use opentitanlib_core::regex;
+pub(crate) use opentitanlib_core::util;
