+++ sw/host/opentitanlib/src/image_lib.rs
@@ -1,0 +1,13 @@
+// Copyright lowRISC contributors (OpenTitan project).
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod image;
+
+// Re-export core modules for internal use to avoid changing crate:: imports
+pub(crate) use opentitanlib_core::crypto;
+pub(crate) use opentitanlib_core::util;
+pub(crate) use opentitanlib_core::with_unknown;
+
+// Re-export chip modules for internal use
+pub(crate) use opentitanlib_chip::chip;
