+++ sw/host/opentitanlib/src/io/i2c.rs
@@ -10,7 +10,6 @@ use std::time::Duration;
 use thiserror::Error;
 
 use super::gpio;
-use crate::app::TransportWrapper;
 use crate::impl_serializable_error;
 use crate::transport::TransportError;
 use crate::util::parse_int::ParseInt;
@@ -32,23 +31,6 @@ pub struct I2cParams {
         value_parser = u8::from_str
     )]
     pub addr: Option<u8>,
-}
-
-impl I2cParams {
-    pub fn create(
-        &self,
-        transport: &TransportWrapper,
-        default_instance: &str,
-    ) -> Result<Rc<dyn Bus>> {
-        let i2c = transport.i2c(self.bus.as_deref().unwrap_or(default_instance))?;
-        if let Some(speed) = self.speed {
-            i2c.set_max_speed(speed)?;
-        }
-        if let Some(addr) = self.addr {
-            i2c.set_default_address(addr)?;
-        }
-        Ok(i2c)
-    }
 }
 
 /// Errors related to the I2C interface and I2C transactions.
