+++ sw/host/opentitanlib/src/io/mod.rs
@@ -2,9 +2,12 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod bootstrap_types;
+pub mod capabilities;
 pub mod console;
 pub mod eeprom;
 pub mod emu;
+pub mod errors;
 pub mod gpio;
 pub mod i2c;
 pub mod ioexpander;
