+++ sw/host/opentitanlib/src/io/spi.rs
@@ -10,7 +10,6 @@ use std::str::FromStr;
 use thiserror::Error;
 
 use super::{eeprom, gpio};
-use crate::app::TransportWrapper;
 use crate::impl_serializable_error;
 use crate::transport::TransportError;
 use crate::util::voltage::Voltage;
@@ -36,35 +35,6 @@ pub struct SpiParams {
     /// SPI polarity/phase mode.
     #[arg(long)]
     pub mode: Option<TransferMode>,
-}
-
-impl SpiParams {
-    pub fn create(
-        &self,
-        transport: &TransportWrapper,
-        default_instance: &str,
-    ) -> Result<Rc<dyn Target>> {
-        let spi = transport.spi(self.bus.as_deref().unwrap_or(default_instance))?;
-        if let Some(ref cs) = self.chip_select {
-            spi.set_pins(
-                None,
-                None,
-                None,
-                Some(&transport.gpio_pin(cs.as_str())?),
-                None,
-            )?;
-        }
-        if let Some(speed) = self.speed {
-            spi.set_max_speed(speed)?;
-        }
-        if let Some(voltage) = self.voltage {
-            spi.set_voltage(voltage)?;
-        }
-        if let Some(mode) = self.mode {
-            spi.set_transfer_mode(mode)?;
-        }
-        Ok(spi)
-    }
 }
 
 /// Errors related to the SPI interface and SPI transactions.
