+++ sw/host/opentitanlib/src/test_utils_lib.rs
@@ -1,0 +1,21 @@
+// Copyright lowRISC contributors (OpenTitan project).
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod test_utils;
+
+// Re-export core modules for internal use
+pub(crate) use opentitanlib_app::transport;
+pub(crate) use opentitanlib_core::impl_serializable_error;
+pub(crate) use opentitanlib_core::io;
+pub(crate) use opentitanlib_core::regex;
+pub(crate) use opentitanlib_core::util;
+
+// Re-export app modules for internal use
+pub(crate) use opentitanlib_app::app;
+
+// Re-export backend modules for internal use
+pub(crate) use opentitanlib_backend::backend;
+
+// Re-export protocols modules for internal use
+pub(crate) use opentitanlib_protocols::bootstrap;
