+++ sw/host/opentitantool/src/command/lc.rs
@@ -98,9 +98,8 @@ impl CommandDispatch for LcStateRead {
         }
 
         // Spawn an OpenOCD process and connect to the LC JTAG TAP.
-        let mut jtag = self
-            .jtag_params
-            .create(transport)?
+        let mut jtag = transport
+            .create_jtag(&self.jtag_params)?
             .connect(JtagTap::LcTap)?;
 
         // Read and decode the LC state.
@@ -137,9 +136,8 @@ impl CommandDispatch for LcRegRead {
         transport.pin_strapping("PINMUX_TAP_LC")?.apply()?;
 
         // Spawn an OpenOCD process and connect to the LC JTAG TAP.
-        let mut jtag = self
-            .jtag_params
-            .create(transport)?
+        let mut jtag = transport
+            .create_jtag(&self.jtag_params)?
             .connect(JtagTap::LcTap)?;
 
         // Read the CSR
@@ -170,9 +168,8 @@ impl CommandDispatch for LcDeviceIdRead {
         transport.reset(UartRx::Clear)?;
 
         // Spawn an OpenOCD process and connect to the LC JTAG TAP.
-        let mut jtag = self
-            .jtag_params
-            .create(transport)?
+        let mut jtag = transport
+            .create_jtag(&self.jtag_params)?
             .connect(JtagTap::LcTap)?;
 
         // Read and concatenate device ID registers.
@@ -225,9 +222,8 @@ impl CommandDispatch for RawUnlock {
         transport.reset(UartRx::Clear)?;
 
         // Spawn an OpenOCD process and connect to the LC JTAG TAP.
-        let mut jtag = self
-            .jtag_params
-            .create(transport)?
+        let mut jtag = transport
+            .create_jtag(&self.jtag_params)?
             .connect(JtagTap::LcTap)?;
 
         let token = parse_token_str(self.token.as_str())?;
@@ -244,9 +240,8 @@ impl CommandDispatch for RawUnlock {
             /*reset_tap_straps=*/ Some(JtagTap::LcTap),
         )?;
 
-        jtag = self
-            .jtag_params
-            .create(transport)?
+        jtag = transport
+            .create_jtag(&self.jtag_params)?
             .connect(JtagTap::LcTap)?;
 
         // Read and decode the LC state.
@@ -292,9 +287,8 @@ impl CommandDispatch for Transition {
         transport.reset(UartRx::Clear)?;
 
         // Spawn an OpenOCD process and connect to the LC JTAG TAP.
-        let mut jtag = self
-            .jtag_params
-            .create(transport)?
+        let mut jtag = transport
+            .create_jtag(&self.jtag_params)?
             .connect(JtagTap::LcTap)?;
 
         let strap_name = if self.use_rma_bootstrap {
@@ -325,9 +319,8 @@ impl CommandDispatch for Transition {
             /*reset_tap_straps=*/ Some(JtagTap::LcTap),
         )?;
 
-        jtag = self
-            .jtag_params
-            .create(transport)?
+        jtag = transport
+            .create_jtag(&self.jtag_params)?
             .connect(JtagTap::LcTap)?;
 
         // Read and decode the LC state.
@@ -374,9 +367,8 @@ impl CommandDispatch for Status {
         transport.reset(UartRx::Clear)?;
 
         // Spawn an OpenOCD process, connect to the LC JTAG TAP, read register, and shutdown OpenOCD.
-        let mut jtag = self
-            .jtag_params
-            .create(transport)?
+        let mut jtag = transport
+            .create_jtag(&self.jtag_params)?
             .connect(JtagTap::LcTap)?;
         let status = jtag.read_lc_ctrl_reg(&LcCtrlReg::Status)?;
         jtag.disconnect()?;
@@ -421,9 +413,8 @@ impl CommandDispatch for TransitionCount {
         transport.reset(UartRx::Clear)?;
 
         // Spawn an OpenOCD process, connect to the LC JTAG TAP, read register, and shutdown OpenOCD.
-        let mut jtag = self
-            .jtag_params
-            .create(transport)?
+        let mut jtag = transport
+            .create_jtag(&self.jtag_params)?
             .connect(JtagTap::LcTap)?;
         let transition_count = jtag.read_lc_ctrl_reg(&LcCtrlReg::LcTransitionCnt)?;
         jtag.disconnect()?;
@@ -454,9 +445,8 @@ impl CommandDispatch for VolatileRawUnlock {
         transport.reset(UartRx::Clear)?;
 
         // Spawn an OpenOCD process and connect to the LC JTAG TAP.
-        let mut jtag = self
-            .jtag_params
-            .create(transport)?
+        let mut jtag = transport
+            .create_jtag(&self.jtag_params)?
             .connect(JtagTap::LcTap)?;
 
         let token = parse_token_str(self.token.as_str())?;
@@ -477,9 +467,8 @@ impl CommandDispatch for VolatileRawUnlock {
             false,
         )?;
 
-        jtag = self
-            .jtag_params
-            .create(transport)?
+        jtag = transport
+            .create_jtag(&self.jtag_params)?
             .connect(JtagTap::LcTap)?;
 
         // Read and decode the LC state.
